requirement.txt
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
numpy-stl==3.1.1
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-utils==3.8.2
requests==2.32.3
scooby==0.10.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
wfastcgi==3.0.0
